+++ clipspyx/__init__.py
@@ -66,6 +66,7 @@ if CLIPS_MAJOR >= 7:
     from clipspyx.tables import Deftable
     from clipspyx.async_goals import (
         GoalHandlerError, enable_goal_handlers, disable_goal_handlers,
+        AsyncRunner,
         AFTER, AT, EVERY,
     )
     from clipspyx.dsl.timer import TimerEvent
@@ -73,5 +74,6 @@ if CLIPS_MAJOR >= 7:
                          'GoalHandlerError',
                          'enable_goal_handlers',
                          'disable_goal_handlers',
+                         'AsyncRunner',
                          'AFTER', 'AT', 'EVERY',
                          'TimerEvent')